apps/provisioning-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "provisioning-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "temporalio>=1.7.0",
    "structlog>=24.2.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
provisioning-agent = "src.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]
